requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
djangorestframework==3.16.0
idna==3.10
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
pre-commit==4.2.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
dj-database-url==2.1.0
pytest==7.4.3
pytest-django==4.7.0
pytest-mock==3.12.0
responses==0.24.1
